+++ src/resources/views/portfolios/view.blade.php
@@ -12,8 +12,8 @@
         <div class="col-md-12">
             <div class="nav-controls text-right">
                 <div class="btn-group" role="group">
-                @if (count($portfolios) > 0)
-                <a href="" class="btn btn-default btn-sm disabled btn-text">{{ $portfolios->firstItem() . ' to ' . $portfolios->lastItem() . ' of ' . $portfolios->total() }}</a>
+                @if (count($models) > 0)
+                <a href="" class="btn btn-default btn-sm disabled btn-text">{{ $models->firstItem() . ' to ' . $models->lastItem() . ' of ' . $models->total() }}</a>
                 @endif
                 {!! HTML::link('admin/portfolios/create', Lang::get('redminportal::buttons.create_new'), array('class' => 'btn btn-primary btn-sm')) !!}
             </div>
@@ -21,20 +21,21 @@
         </div>
     </div>
     
-    @if (count($portfolios) > 0)
+    @if (count($models) > 0)
         <table class='table table-striped table-bordered table-condensed'>
             <thead>
                 <tr>
-                    <th>{{ Lang::get('redminportal::forms.name') }}</th>
-                    <th>{{ Lang::get('redminportal::forms.category') }}</th>
-                    <th>{{ Lang::get('redminportal::forms.summary') }}</th>
-                    <th>{{ Lang::get('redminportal::forms.active') }}</th>
-                    <th>{{ Lang::get('redminportal::forms.updated') }}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/portfolios', 'name', $sortBy, $orderBy) !!}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/portfolios', 'category_name', $sortBy, $orderBy, trans('redminportal::forms.category')) !!}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/portfolios', 'short_description', $sortBy, $orderBy, trans('redminportal::forms.summary')) !!}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/portfolios', 'active', $sortBy, $orderBy) !!}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/portfolios', 'created_at', $sortBy, $orderBy) !!}</th>
+                    <th>{!! Redminportal::html()->sorter('admin/portfolios', 'updated_at', $sortBy, $orderBy) !!}</th>
                     <th></th>
                 </tr>
             </thead>
             <tbody>
-            @foreach ($portfolios as $portfolio)
+            @foreach ($models as $portfolio)
                 <tr>
                     <td>{{ $portfolio->name }}</td>
                     <td>{{ $portfolio->category->name }}</td>
@@ -46,7 +47,8 @@
                             <span class="label label-danger"><span class='glyphicon glyphicon-remove'></span></span>
                         @endif
                     </td>
-                    <td>{{ date('d-M-y', strtotime($portfolio->updated_at)) }}</td>
+                    <td>{{ date('d/m/y h:i A', strtotime($portfolio->created_at)) }}</td>
+                    <td>{{ date('d/m/y h:i A', strtotime($portfolio->updated_at)) }}</td>
                     <td class="table-actions text-right">
                         <div class="btn-group">
                             <button type="button" class="btn btn-link dropdown-toggle" data-toggle="dropdown">
@@ -69,7 +71,7 @@
             </tbody>
         </table>
         <div class="text-center">
-        {!! $portfolios->render() !!}
+        {!! $models->render() !!}
         </div>
     @else
         <div class="alert alert-info">{{ Lang::get('redminportal::messages.no_portfolio_found') }}</div>